% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cases_per_gen.R
\name{mod_cases_per_gen_ui}
\alias{mod_cases_per_gen_ui}
\alias{mod_cases_per_gen_server}
\title{mod_cases_per_gen_ui and mod_cases_per_gen_server}
\usage{
mod_cases_per_gen_ui(id)

mod_cases_per_gen_server(input, output, session, sliced_sims)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Extract and summarise the number of cases per generation
}
\keyword{internal}
