% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_effective_r0.R
\name{mod_effective_r0_ui}
\alias{mod_effective_r0_ui}
\alias{mod_effective_r0_server}
\title{mod_effective_r0_ui and mod_effective_r0_server}
\usage{
mod_effective_r0_ui(id)

mod_effective_r0_server(input, output, session, sliced_sims)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{sliced_sims}{A single timepoint from each model simulation}
}
\description{
Extract and summarise the effective R0
}
\keyword{internal}
