% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_extinct_prob.R
\name{mod_extinct_prob_ui}
\alias{mod_extinct_prob_ui}
\alias{mod_extinct_prob_server}
\title{mod_extinct_prob_ui and mod_extinct_prob_server}
\usage{
mod_extinct_prob_ui(id)

mod_extinct_prob_server(input, output, session, sliced_sims, setup)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{sliced_sims}{dataframe of the first time point of model simulations}

\item{setup}{A list of inputs for the model}
}
\description{
A shiny Module.
}
\keyword{internal}
