% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_slice_time_point.R
\name{mod_slice_time_point_ui}
\alias{mod_slice_time_point_ui}
\alias{mod_slice_time_point_server}
\title{mod_slice_time_point_ui and mod_slice_time_point_server}
\usage{
mod_slice_time_point_ui(id)

mod_slice_time_point_server(input, output, session, sims)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{sims}{dataframe of model simulations}
}
\description{
Slice out a single time point from each model simulation
}
\keyword{internal}
