% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generation_size.R
\name{plot_generation_size}
\alias{plot_generation_size}
\title{Plot cases per generation}
\usage{
plot_generation_size(generation_sizes, log = FALSE)
}
\arguments{
\item{generation_sizes}{A dataframe containing cases per generation}

\item{log}{Logical, defaults to FALSE. Should the plot be shown logged on the y axis.}
}
\value{
A `ggplot2` object plotting cases per generation
}
\description{
Plot cases per generation
}
\examples{


}
