% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traces.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot outbreak traces}
\usage{
plot_traces(sims, log = FALSE)
}
\arguments{
\item{sims}{A dataframe produced by `ringbp::scenarios_sims`.}
}
\value{
A `ggplot2` plot of outbreak traces
}
\description{
Plot outbreak traces
}
\examples{


}
