% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_extinct_prob.R
\name{mod_extinct_prob_ui}
\alias{mod_extinct_prob_ui}
\alias{mod_extinct_prob_server}
\title{mod_extinct_prob_ui and mod_extinct_prob_server}
\usage{
mod_extinct_prob_ui(id)

mod_extinct_prob_server(input, output, session, sims)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{sims}{dataframe of model simulations}
}
\description{
A shiny Module.
}
\keyword{internal}
